Move EmptyRequestBodyReader to its own file.
