+++ src/Microsoft.AspNetCore.Server.Kestrel.Core/Internal/Http/RequestBodyReader.cs
@@ -157,25 +157,4 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
             }
         }
     }
-
-    public class EmptyRequestBodyReader : IRequestBodyReader
-    {
-        public static readonly IRequestBodyReader Instance = new EmptyRequestBodyReader();
-
-        public Task ConsumeAsync(CancellationToken cancellationToken = default(CancellationToken))
-            => TaskCache.CompletedTask;
-
-        public Task CopyToAsync(Stream destination, CancellationToken cancellationToken = default(CancellationToken))
-            => TaskCache.CompletedTask;
-
-        public Task<int> ReadAsync(ArraySegment<byte> buffer, CancellationToken cancellationToken = default(CancellationToken))
-            => TaskCache<int>.DefaultCompletedTask;
-
-        public void Reset()
-        {
-        }
-
-        public Task StartAsync(MessageBody messageBody, CancellationToken cancellationToken = default(CancellationToken))
-            => TaskCache.CompletedTask;
-    }
 }
